elf: Do not process invalid tunable format
Tunables string with with more than one '=' on its definition are
parsed and enabled, and any subsequent '=' are ignored.  It means
that tunables in the form 'tunable=tunable=value' or
'tunable=value=value' are handled 'tunable=value'.  This inputs are
most likely user input error, which should not be accepted.

Checked on x86_64-linux-gnu.